+++ packages/gateway/src/graphql/schema.d.ts
@@ -88,6 +88,7 @@ export interface GQLMutation {
   markEventAsDuplicate: string
   confirmRegistration: string
   rejectRegistration: string
+  upsertRegistrationIdentifier: string
   createOrUpdateUser: GQLUser
   activateUser?: string
   changePassword?: string
@@ -601,7 +602,6 @@ export interface GQLReinstated {
 
 export interface GQLConfirmRegistrationInput {
   registrationNumber: string
-  error?: string
   identifiers?: Array<GQLIdentifierInput>
 }
 
@@ -2633,6 +2633,7 @@ export interface GQLMutationTypeResolver<TParent = any> {
   markEventAsDuplicate?: MutationToMarkEventAsDuplicateResolver<TParent>
   confirmRegistration?: MutationToConfirmRegistrationResolver<TParent>
   rejectRegistration?: MutationToRejectRegistrationResolver<TParent>
+  upsertRegistrationIdentifier?: MutationToUpsertRegistrationIdentifierResolver<TParent>
   createOrUpdateUser?: MutationToCreateOrUpdateUserResolver<TParent>
   activateUser?: MutationToActivateUserResolver<TParent>
   changePassword?: MutationToChangePasswordResolver<TParent>
@@ -3204,6 +3205,23 @@ export interface MutationToRejectRegistrationResolver<
   (
     parent: TParent,
     args: MutationToRejectRegistrationArgs,
+    context: Context,
+    info: GraphQLResolveInfo
+  ): TResult
+}
+
+export interface MutationToUpsertRegistrationIdentifierArgs {
+  id: string
+  identifierType: string
+  identifierValue: string
+}
+export interface MutationToUpsertRegistrationIdentifierResolver<
+  TParent = any,
+  TResult = any
+> {
+  (
+    parent: TParent,
+    args: MutationToUpsertRegistrationIdentifierArgs,
     context: Context,
     info: GraphQLResolveInfo
   ): TResult